+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added type `ChannelAssignmentAttribute` for vertex attribute shader dependencies.
+* Added method `attribute` to `ChannelAssignment`.
+
+### Changed
+* Renamed `Mesh.unk_mesh_index1` to `Mesh.index_buffer_index2` to better reflect in game data.
+
 ## 0.6.0 - 2024-06-09
 ### Added
 * Added fields `unk_mesh_index1`, `ext_mesh_index`, and `base_mesh_index` to `Mesh`.